feat(web): remove org scope from task runs and schedules
