+++ ORGANISATION_REMOVAL_TASKS.md
@@ -634,7 +634,7 @@ outside the task-run-heavy group operations view.
 
 ## Task 10 - Web Task Runs And Schedule Conversion
 
-Status: Not started
+Status: In progress
 
 Completed by:
