+++ apps/web/app/(dashboard)/tasks/page.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from 'next'
 import { redirect } from 'next/navigation'
 import { getRequiredSession } from '@/lib/auth/session'
 import { canAccessTooling } from '@/lib/auth/tooling'
+import { resolveCurrentActionScope } from '@/lib/actions/action-scope'
 import { listSchedules } from '@/lib/actions/task-schedules'
 import { SchedulesClient } from './schedules-client'
 
@@ -12,13 +13,11 @@ export const metadata: Metadata = {
 export default async function ScheduledTasksPage() {
   const session = await getRequiredSession()
   if (!canAccessTooling(session.user)) redirect('/dashboard')
-  const orgId = session.user.organisationId!
-
-  const schedules = await listSchedules(orgId)
+  const currentScope = resolveCurrentActionScope(session)
+  const schedules = await listSchedules(currentScope)
 
   return (
     <SchedulesClient
-      orgId={orgId}
       userRole={session.user.role}
       initialSchedules={schedules}
     />
